+++ module/auth/src/main/kotlin/io/github/siyual_park/auth/configuration/AuthenticatorConfiguration.kt
@@ -1,13 +1,11 @@
 package io.github.siyual_park.auth.configuration
 
-import io.github.siyual_park.auth.annotation.AuthenticateMapping
-import io.github.siyual_park.auth.domain.authenticator.AuthenticateFilter
+import io.github.siyual_park.auth.domain.authenticator.AuthenticateFilterFactory
+import io.github.siyual_park.auth.domain.authenticator.AuthenticateMapping
 import io.github.siyual_park.auth.domain.authenticator.AuthenticateProcessor
 import io.github.siyual_park.auth.domain.authenticator.Authenticator
 import io.github.siyual_park.auth.domain.authenticator.AuthorizationAuthenticateProcessor
 import io.github.siyual_park.auth.domain.authenticator.AuthorizationProcessor
-import io.github.siyual_park.auth.domain.authenticator.MatchTypeAuthenticateFilter
-import org.springframework.beans.BeansException
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.ApplicationContext
 import org.springframework.context.annotation.Configuration
@@ -17,27 +15,12 @@ class AuthenticatorConfiguration(
     private val applicationContext: ApplicationContext
 ) {
     @Autowired(required = true)
-    fun configAuthenticatorManager(authenticator: Authenticator) {
+    fun configAuthenticatorManager(authenticator: Authenticator, filterFactory: AuthenticateFilterFactory) {
         applicationContext.getBeansOfType(AuthenticateProcessor::class.java).values.forEach {
             it.javaClass.annotations.filter { it is AuthenticateMapping }
                 .forEach { annotation ->
                     if (annotation !is AuthenticateMapping) return@forEach
-                    val filterBeen = try {
-                        applicationContext.getBean(annotation.filterBy.java)
-                    } catch (e: BeansException) {
-                        null
-                    }
-                    val filter = when (filterBeen) {
-                        null -> {
-                            MatchTypeAuthenticateFilter(annotation.filterBy.java)
-                        }
-                        is AuthenticateFilter -> {
-                            filterBeen
-                        }
-                        else -> {
-                            return@forEach
-                        }
-                    }
+                    val filter = filterFactory.create(annotation) ?: return@forEach
                     authenticator.register(filter, it)
                 }
         }
